Initial commit
This is the initial commit for the library bandsaw. It includes the
fundamental classes with tests, the build scripts and the structure of
the documentation without content.
The current revision should be enough to automate builds and deployment
of the package and documentation, but is not meant for an official
release.