Fix: Removed Protocol base class to be backwards compatible
The typing.Protocol type is only available since Python 3.8. Therefore
we imported it from typing_extensions which has the drawback, that
this package is not part of the standard library and needs to be
installed additionally. Since the base class Protocol isn't really
needed but is just some convenience marker for static analyzers, we
remove it completely so that the module is compatible with all standard
python versions 3.6+. If we later decide to support only 3.8+ we can
easily add the Protocol base again.