Add function to create executable to run session
This commit adds a new function, that allows to package the code
necessary to continue a session in a different interpreter. This
package is a zip archive with a __main__ module that loads a
session from file, continues it and writes the resulting session
to a different file.
We use this now when running a task in a subprocess, which means
we no longer need to take care that the python path of the parent
process is used for the child process, too.
The runner archive will be especially needed when we transfer the
session to another machine.