Moved distribution archive creation to own module
This commit moves the code for creating a distribution archive to its
own module. Additionally, we now get the distribution archive from a
session, which will allow us to cache the archives and create them
lazily later on. To make the distribution archive configurable, we
now add a new method to Configuration to define additional modules
that should be added to the archive, too.