Create new class DistributionArchive
This commit introduces a new class DistributionArchive which represents
the concept of code that needs to be distributed if we want to run tasks
in distributed fashion. Instead of creating the distribution archive
explicitly in our Session, we now have a new function
`get_distribution_archive(configuration)` which returns a distribution
archive for a specific bandsaw configuration. Having this function will
later be used for caching archives.
Since not all configurations or advices will need a distribution archive
we now create the archive lazily as soon as someone requests its path.