Add run_id to uniquely identify workflow runs
We introduce a new free function get_run_id() that returns a unique id
that identifies the individual workflow run. This allows to track task
executions that were calculated in the same run but differentiate them
from the same executions from the same workflow but from a different
run.
The run_id has to be set explicitly using set_run_id() when the workflow
is distributed to other python interpreters.