Reduce length of identifiers and make it configurable
Identifiers were previously 64 characters long, which makes them hard to
read. Most of the time, we don't need the full length of the identifier
to prevent collisions, so we limit the length now to 16 characters. To
be safe, we allow to change the length of the identifier using the
'BANDSAW_ID_LENGTH' environment variable.