Fix: Specifying chain on task was broken
Trying to specify only the chain or the configuration in a task
decorator was broken since they weren't recognized as being part of the
task kwargs. We now fix the check to explicitly check for both keyword
arguments and add two additional unit tests to make sure this can't
break again.