Add attachments to session
Often advices want to attach some additional data to a session, e.g.
metrics. We now add some new attribute `attachments` to the `Session`
class that allows to attach arbitrary file to the session. When a
session is saved or restored, these additional attachments are added to
the session archive and transferred alongside.