Add configuration for temporary directory
Often advices need to store temporary files, e.g. when copying over a
session to a different machine or storing temporary state. This leads to
multiple advices using their own temporary directory, but since an
advice doesn't know when its no longer used, these directories are most
of the time not proper cleaned up. We now introduce a new temporary
directory to the configuration (and later the session) that can be used
instead and is guaranteed to be removed when the python interpreter
exits.