LoggingAdvice stores now the log as attachment in session
This commit extends the LoggingAdvice to store the log items in a log
file that is later added to the session as an attachment 'session.log'.
This is done by adding a new FileHandler in `before()` and removing it
later in the `after()` callback. The file handler writes the log file
to the session temporary directory. As a default formatter we use a new
formatter that writes the log items as JSON.