Add function to summarize info about values
As soon as we implement some kind of user interface for bandsaw, that
gives users more insight into the structure of a workflow, we need to
give additional information about the values that are used as arguments
and results. This commit introduces a first function `value_info()` that
creates an info object in form of a dict that gives information about
type, representation and size of a value.