Add TrackerExtension to track sessions, tasks and executions
This commit adds a new TrackerExtension which allows to keep track of
different runs of a workflow and the computed tasks within. For this,
the extension supports different backends that store the information
about the workflow. As a first implementation of a backend, this commit
includes a FileSystemBacked, that writes the information in json format
to a configurable directory.