Unify generation of task id, support version to change id
This commit unifies the way, the task id is derived in all types of
tasks. Additionally, we now support changing the id by setting the
"version" advice parameter. We discovered a problem with the unit tests,
where the module wasn't reloaded after changing the id length. This is
now fixed as well, so that tests are no longer dependend on the order of
their execution.