Use Context class instead of dict() for storing context
We now use our own Context class for storing the session context instead
of the native dict, since we want to make sure, that the context only
changes by adding new values. This enforces, that information stored
later about a session is always a superset of information that was
stored before and now conflicts can come up.